fix: now cane reduce product quantity from cart with patch method instead seprate endpoint

+++ shopping/cart/models.py
@@ -23,7 +23,7 @@ class CartItem(models.Model):
         null=True,
         blank=True,
     )
-    quantity = models.PositiveIntegerField(default=1)
+    quantity = models.SmallIntegerField(default=1)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
